campusstay_backend/requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.23.2
SQLAlchemy==2.0.22
psycopg2-binary>=2.9.9
python-dotenv==1.0.0
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
alembic>=1.13.1